[IMP] helpdesk: improve the helpdesk team kanban and ticket breadcrumb
Purpose of the task is, when some one remove the default filter
at that time ticket kanban view is really a mess so when go though the
helpdesk team always apply the domain of current team instead
of filter.

The breadcrumb should be the name of the helpdesk team instead of name
of the action(tickets) so pass display_name from method to change the
breadcrum when view the ticket from kanban view.

closes odoo/enterprise#6903

Task: 2124377
Closes: #6903
Related: odoo/odoo#40922
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>