[FIX] stock_barcode: put in pack in receipt
- Activate product packing
- Create a PO for a stockable product, validate
- In the barcode interface, go to the created picking
- Set the product quantity
- Click on 'Put in Pack'

An error is raised: 'All the products currently reserved in the picking
are already in a pack...'

In the regular interface, the procedure is to create the move lines
manually, therefore setting a quantity done. However, this is not
possible in the barcode, and `move_line_nosuggest_ids` is always empty.

We introduce a special use case to bypass this in the case of the
barcode view.

opw-2127637

closes odoo/enterprise#6906

X-original-commit: 5f8796d5692b73031023a3700736a5a090e2b061
Related: odoo/odoo#40928
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>