[FIX] web_gantt: illegal attribute
The method to postprocess access rights add some default attribute
on the main node of the view to allow/disallow creation/edition of
records through the view globally.

In the case of the Gantt view, it added a `delete` attribute that is,
in fact, not in the correct schema of the gantt view (check the rng
file if you don't believe me).

This meant that in some cases, a perfectly valid view didn't pass
validation because this method added illegal attributes.

closes odoo/enterprise#6971

X-original-commit: c13a3995ae5fbbe44ca0e46d148daddceb5dbc69
Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>