[FIX] web_gantt: do not attempt to load control panel in dashboard
Insatll dashboard and another app with gantt view (i.e. planning).
Go to planning, check out the gantt view and add to dashboard
(favorites>Add to Dashboard).
Go to dashboard

Javascript traceback will popup. Gantt model is initialized without the
control panel, so the buttons are not initialized and there is an
attempt of resize that cause the issue.

Switch to operate on buttons only when the control panel is defined.

opw-2128190

closes odoo/enterprise#6967

X-original-commit: 6657e992322034cf5f6d6965da5b05fae2f5323f
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>