[REF] quality, *: multiple products on a quality point.
Currently the configuration of quality point is not optimal. For example
if you want to check 5 differents product during your receipt then
you have to create 5 points, one for each different product. Same
behavior for the product's variants.

Ease the configuration by using a m2m instead of a m2o on the quality
points. 3 configurations are allowed:
- No product templates nor variants defined on the point. Then the point
is applied to all products
- A product template is defined but not its variants. Then the point is
applied to all variants.
- Variants are defined. The point is applied to specified variants.

Technicaly the quality check creation for mrp.production has been moved
on stock.move (to be coherent with picking). Some methods has been
modify as helper in order to reduce duplicate code.

Task: 31065