[IMP] timesheet_grid: add condition to filter "My team"
TaskID: 2088586

Before this commit, the filter searches the analytic lines where the employee
has the current user as timesheet responsible.
Now, the filter searches the analytic lines where the owner is
the current user or the employee has the current as timesheet
responsible or the manager,
or the owner of the project of the timesheet is the current user.