[FIX] l10n_mx_edi_landing: Customs number now is assigned.
The method ``_run_fifo()`` now is in the product, not in the move like
in the last version. And extra, now is not send the ``move`` as
parameter, to know which is the move where must be assigned the fifo
candidate.

Now, to get the move origin, was moved to ``_create_out_svl``, to assign
the entry correctly.

The tests was improved.

closes odoo/enterprise#7123

X-original-commit: d62cab1d0440cf6d3201f06f42e7c318a61cc457
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>