[ADD] {account,l10n_lu,l10n_no}_saft: add SAF-T export
Task 2150409

SAF-T (FAIA for Luxembourg) is a format for exporting the journal
entries along with other informations about the related products and
partners.
This format becomes required by the government in multiple countries.

Because therecan be some changes specific to each country, we made a
base module that should be inherited by each implementation of SAF-T.

closes odoo/enterprise#10032

X-original-commit: aa5441a2206d6d92e23649288eed328a7ea9ed33
Signed-off-by: Josse Colpaert <jco@openerp.com>
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>