[FIX] timesheet_grid: no reordered records when we click on timer button
TaskID: 2088586

Before this commit, when the user click on the timer button, we render all the view. The order of the timesheets change because the default order is "date desc, id desc".

We don't want to reorder this because it could be confusing for the users.

To revolve this, the updateState in renderer is override to update only the grids, rows and cols property of the state when the user click on the timer button.
For the other cases, the updateState method call his parent.