[FIX] account_accountant: reconcilation partial amount nice locale
If you reconcile for example a bank statement of 200,00€ with a
customer invoice of 300,00€, you get a pen icon next to the 300,00€
and if you click on it you have an editing cell with:

200.00

But if you validate this, this will be parse in the current language
number formatting so if you have . as thousands separator, this number
will become 20000 which is not expected.

With this changeset, we display the number in the current lang format.

note: with this fix, number of decimals can be different than 2

note: forward-port of odoo/odoo#41410 saas-12.3 adapted for saas-13.1

opw-2150396
closes #7078

closes odoo/enterprise#7151

X-original-commit: f7e00a8e06652af3ce2b72c083a1fc0194bebd4e
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>