[IMP] inter_company_rules: add rule in company for inter_company_user_id
This commit is done to show the superuser in the m2o dropdown so that we
can change from other users to superuser whenever required and changed lable
to create as.

Previously, only default value was shown as superuser but
if we remove it and change it to other users than we were not able
to set it back to the superuser.

Task - 1964526

closes odoo/enterprise#4244

Signed-off-by: Damien Bouvy (dbo) <dbo@odoo.com>