[IMP] hr_timesheet: the form view can edit only unit_amount field of validated timesheets
TaskID: 2088586

Before this commit, in the kanban for timesheets, the user can click on
a timesheet to go to the form view of this timesheet. Even if this
timesheet is validated, the user can edit all available fields in the form
view.

Now, if the timesheet selected by the user in the kanban view is
validated, he may only modify the unit_amount field of this timesheet.