[ADD] l10n_co_edi_ubl_2_1, product_unspsc: New EDI format Colombia + UNSPSC codes
This is a forward port of the original modules in v12 where
we needed to adapt to the new account.invoice -> account.move model
Also for the UNSPSC module, a change of view was required.

From December 2019, the DIAN (Colombian government) asks for
a different format in UBL 2.1 for the electronic invoices to send to them.
This new format also needs a lot more information.  This is why
we added a module in v12 with the new format such that when you
install it, it will start sending in the new format.  As we pass through
Carvajal, the file is itself not ubl2.1 but still the updated Carvajal format.

The goal was not to refactor the old module, but to make the new format work.

- We load the EDI codes for the taxes from the templates with a post_init_hook
(in order not to have to configure them manually)
- When downloading the answer from a file, we take the CUDE/CUFE from the file
and put it in a field as we need it for sending credit note/debit note
- The new module depends on base_address_city as we need to send codes for
the cities, states, ...
- The new module also depends on the new product_unspsc to load the UNSPSC
codes in case you don't have a barcode.
- Payment options are added, but it is just a field on the invoice (as we only
send customer invoices)
- Also for the debit note, we keep the link with the original invoice in a
separate field.
- We added a field in the return wizard, so when posting the credit immediately
from the wizard, it will contain the correct reason.

closes odoo/enterprise#8086

Task: 2152432
X-original-commit: d46cceef8c594b9056d0115edb7169e207a5986f
Signed-off-by: Josse Colpaert <jco@openerp.com>