[FIX] l10n_co_edi: use the legal status to determine the invoice status
There are two invoice statuses in the response:

- status: status of Carvajal (invoice processing 3rd party)
- legalStatus: status of DIAN (government)

All combinations of the statuses happen, even though it seems weird at
first (not accepted for Carvajal, but accepted according to the
government).

The best way to handle all cases is to look at the legal status as
soon as Carvajal finishes processing it. This is the legally binding
status.

opw-2154293

closes odoo/enterprise#7221

X-original-commit: 4e1b5c99bc0adfe88be79f0fde8c5ace0c70ab66
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>