[FIX] web_mobile: correctly destroy datepicker
In Odoo app, we don't need to instantiate a datepicker
because the native component is used. It's why we destroy
it by overriding "start" method.
But before this commit, it wasn't the case as the wrong
element was taken.
Note that it didn't appear because "readonly" was set.

Now, we destroy it on the same node that it was instantiated:
this.$el (parent node of this.$input).

This commit also fix a bug introduced by the following commit:
https://github.com/odoo/odoo/commit/5481c429efb50be8d8617e4c17b04999c5a7fce8

It seems that it only occurs in 12.0 because events handlers
are different between in each version; in our case:
'change.datetimepicker': 'changeDatetime', is called but this
method is never called in future versions.
See odoo/odoo@ac85176ce81427ba6f55b59e28c1ff6e7e080c32

Steps to reproduce:
- Go to Sales
- Click on "create" (quotation)
- The following message will appear:
"The record has been modified, your changes will be discarded.
Do you want to proceed?"

Then hasChanged is true and the datetime field is considered as
dirty. But now the datepicker is correctly destroyed and calls to
"changeDatetime" are correctly skipped by calling "return;" thanks
to "__libInput".

opw-2154200

closes odoo/enterprise#7301

X-original-commit: 015a9f05b2da9c48fa8f916157a80542373d74dc
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>