[FIX] documents: prevents wrong mimetype on write
This fix prevents the ir.attachment mimetype guessing mechanism
to take over the mimetype attribution when explicitly giving the
mimetype.

How to reproduce: write on a document with a `datas` and a
`mimetype` of type 'text/plain' or 'application/zip'. Can
be done with the "replace file" feature of the document
inspector.

closes odoo/enterprise#7295

Task: 2156396
X-original-commit: dbd50eade852abef88d591f78795e8c37c67346d
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>