[FIX] sale_account_taxcloud: downpayment taxes counted twice
This makes a "fix" from the "workaround" from:
https://github.com/odoo/odoo/issues/20986#issuecomment-345975748

A downpayment is created as a positive invoice, and then subtracted from
the regular invoice. Since negative lines are ignored, it means
downpayment are counted twice.

The workaround is to set the deposit product TIC category as "Gift card"
since no tax is applied on such products.
So in TaxCloud we automatically check that the correct category is set
on the deposit product, so that no configuration is needed to make it
work properly.

opw 2079677

closes odoo/enterprise#7330

X-original-commit: 5faa6d54be32658f24b3dc78b4ac8956d75682a3
Signed-off-by: Nans Lefebvre (len) <len@odoo.com>