[IMP] hr_appraisal : improve generation of appraisal dates.
The hr.employee module now has a last_appraisal_date and a
next_appraisal_date field to better manage appraisals.

The next_appraisal_date is generated upon the creation of an
employee and is the creation date + the maximum duration between appraisals.
It is editable so that it can be corrected when the employee form
is created before the employee starts working in the company.

Task-2154738

closes odoo/enterprise#7398

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>