[IMP] account_batch_payment: introduce a new API to report report batch validation errors.
We don't just trigger a UserError anymore. Instead, we open a custom wizard listing the different errors that occured and allowing to directly remove from the batch or see all the payments causing each of them. This allows easier use of the batches, since it does not require them to go through the payments in error on by one manually to either remove or fix them.

[IMP] account_sepa, account_sepa_direct_debit: make use of the new batch payment validation error API.

closes odoo/enterprise#2535

Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>