[IMP] industry_fsm_report : improve multicompany management in industry_fsm_report
This adds a company_ids field on worksheet templates to allow the creation of worksheet
templates specific to one or more companies.

This also adds the corresponding domain to project.project and project.task.

It also adds a wizard that can be used to unlink a worksheet from multiple projects and/or tasks at the same time.

Task-2169872