[FIX] web_enterprise: burger menu isn't automatically closed
Before this commit, the user didn't see the result of an action
which was below the burger menu which remained open.
Now we automatically close the burger menu to be able to see it.

Steps to produce:

1. Install and open 'Employees'
2. Open the burger menu
3. Click on 'Preferences'
=> The burger menu is still open

OR

1. Install the Accounting + l10n_be module
2. Open ths burger menu
3. Click on: Configuration -> Payments -> Add a bank account
=> The burger menu is still open

Due to the implementation of the mobile burger menu, we can't
trigger up the 'close_o_burger_menu'.
It's why we use the bus to provide a way to close the burger menu.

opw-2071605

closes odoo/enterprise#7647

X-original-commit: c772557037c120e9213a4943c9870c4558373fed
Related: odoo/odoo#43102
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>