[FIX] mail_enterprise: trigger DOM updated on attachment rendered
Before this commit, when an attachment was rendered and appended, there was no
notification of any kind to allow the list to recompute its size.

Now, when an attachment is rendered in a form, it triggers "DOM_updated" which
will notify the list that it should render itself anew.

Bug noticed on pad http://pad.odoo.com/p/feedback

closes odoo/enterprise#7375

X-original-commit: 9dcfff8ad681056d15deb7440242447436c547fe
Signed-off-by: Julien Mougenot (JUM) <Arcasias@users.noreply.github.com>