[IMP] helpdesk_sale_timesheet : add a create sale order button on helpdesk tickets and batch invoice
'Create Sale Order' Button
==========================

The button is visible on helpdesk tickets that are linked to a project with allow billable False
or allow billable is True and the ticket is also linked to a task.

The function that computes the sale_order is overriden to recompute it when the sale_order is modified on the project
or on the task.

For the first option, the button opens the same wizard as the create sale order button on a project form.
The second option opens the wizard from the project task form.

There is now also a stat button that redirects the user to the sale order. The button is not visible if the
sale order is not set.

The default configuration of a project created from a ticket are allow_timesheets = True, allow_timesheet_timer = True
and allow_billable = True.

This adds demo data to allow a demonstration of the sale order creation from a ticket.

Removal of overdue alert
========================

The overdue alert was confusing if no hours were set on the task
linked to the ticket. It also wasn't useful anymore as the number
of remaining hours can be seen on the kanban view.

Batch Invoice
=============

This adds a server action to allow the suer to batch invoice tickets
where a sale order has been created. The action is only called on
ticket with a sales order. If there's no tickets with a SO, a UserError
indicates that to the user.
We also add a filter to filter tickets that we can invoice.

Task-2157571