[FIX] helpdesk: avoid horizontal scroll on mobile
Steps to reproduce:
1. Open 'Helpdesk' on mobile
2. In dashboard menu, you can scroll x while the content fit
the screen.

This issue occured because of negative margins on
'o_helpdesk_dashboard' class and the computed width is wrong.
In general, using negative margins should be avoided.

So in this commit we remove negative margins and we adapt
padding values to keep the same design.

padding-top: 2px = 10px + (0 - @o-kanban-record-margin)
padding-bottom: 18px = 10px + @o-kanban-record-margin

opw-2071605

closes odoo/enterprise#7419

X-original-commit: 8874d4717760650c10c6a7d99e02975fafdea037
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>