[IMP] stock_barcode: button to add quantity
In the transfers barcode view, adds two buttons on barcode line:
    - One button to increase done quantity one by one;
    - A second button to add remaining quantity (reserved quantity minus
      done quantity). This one is only displayed if the line has
      reserved quantity.
These buttons are hidden if the line has all its quantity done or if the
product is tracked by serial number. For product tracked by lot number,
these buttons are displayed only if the line has a lot name or a lot id.

Also, selects a line when user clicks on it.
The purpose is to make barcode app usable without barcode scanner.

task-2123726