[FIX] marketing_automation: smaller cron batches
Custom fw-port of https://github.com/odoo/enterprise/pull/7035 to
include https://github.com/odoo/enterprise/pull/7531

The batches were sized according to the number of activities, and not
dependent on the number of traces to handle.

If 1 activity has a lot of traces (too many to perform in 1 cron
execution time), it would never commit its progress, and perform the
execution of the first traces every time the cron is called, executing
their action multiple times.

We instead split in batches of 500 traces. There's no guarantee this is
small enough as it can execute any server action, but it's to be aligned
with the MailComposer's batch limit.

We also patch the test that call this method to ensure it doesn't
actually commit the test-data in the database.

closes odoo/enterprise#7560

X-original-commit: dd994e2344a8421f7f4f1ff0062ceb318a562983
Signed-off-by: Denis Vermylen <Icallhimtest@users.noreply.github.com>