[FIX] account_ponto: fix deadlock in some case and no error when refresh too fast
Don't show error when trying to manually refresh and it results in an error saying that we
already initiated a refresh not too long ago.
Prevent Deadlock in some case:
Deadlock could appear in the following case: call to _ponto_fetch that will result in an error
because access token has expired, in that case we generate a new access token (which will write the
new access token to the record) and then we call back the original route again to retry. This new
call can result in en error which we have to log to the record as well as raise it for the user to see
(we log it to the record because it can happen in a cron as well).
Therefore we have a conflict as we already wrote some value on the record for the new access token and
we try to write on the same record within a new cursor to say that its status should be changed to FAILED
hence a deadlock.

closes odoo/enterprise#7577

X-original-commit: 454eb6adb0384849a262fdebcb7491b8bcc1a652
Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>