[FIX] web_gantt: fix traceback when group by date
Bug
===
If group by "date: month" in the tree view,
and then go to the gantt view which has a default
scale equal to "day", a traceback pop.

Fix
===
We check that we do not group by date in reload,
but we also need to do it during initialization.

Task 2168740