[FIX] account_ponto: convert date in correct timezone
Date received from ponto are always datetime in UTC. We convert those datetime to date
in odoo (hence loosing the time part). However some bank don't provide the time of the
transactions, hence the received UTC datetime we have is smth like 2019-01-01T23:00:00.000z
which is converted to a date 2019-01-01 while the correct transaction date should be
2019-01-02.
Therefore we convert the received datetime in the Brussel Timezone before converting it
to a date.

Also remove the last remnant of code used to skip refresh as it is better handled by an error code

closes odoo/enterprise#7645

X-original-commit: 8f5c00564d469f98ee4f7110a9ac30d2a43258d2
Signed-off-by: Cedric Snauwaert (csn) <csn@openerp.com>