[IMP] account: display 'P' as matching number when account.move.line objects are partially reconciled
For this, we introduce a new matching_number field, used in account.move.line's tree views instead of full_reconcile_id, which used to be shown directly.

[IMP] account_accountant: use transfer wizard to reconcile payable with receivable

Modify the 'reconcile' action defined on account.move.line objects so that it displays the transfer wizard (pre-filled with the right values) to transfer the balance from a receivable account to a payable account. This feature is only available if all the lines we're trying to reconcile belong to the same partner, and if they use exactly one receivable account and one payable account.

Was task: 2037318

closes odoo/enterprise#6544

Related: odoo/odoo#39415
Signed-off-by: Laurent Smet <smetl@users.noreply.github.com>