[FIX] mrp_workorder: onchange on product template
- Create a BOM for 'Customizable Desk'
  Do not choose a variant
  Set the routing 'Primary Assembly'
- Click on 'Primary Assembly', then 'Steps'
- Create a step
- Choose as product the 'Customizable Desk'

A 'Product Variant' is set while no variant has a BOM with the routing
selected.

This is confusing for the users: they see a product selected, but if
they remove it or try to change it, the list is empty.

We don't suggest any product, since there is not much added value to do
so.

opw-2165638

closes odoo/enterprise#7684

X-original-commit: 4d472f3c941719b37911207ca1001b67be9b98ef
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>