[IMP] l10n_in_hr_payroll: Several data, view and fields improvement.
Purpose of this task is to move l10n_in _hr_payroll demo data to data
and imrpvoe the view and add several required field.

In this commit ,We moved demo data to data file with some improvement.
And we also add three new fields in indian payroll module and also
display those field on the payslip report.
And updated view so Allowance and Deduction Both parts should be shown
in salary details.

closes odoo/enterprise#8291

Taskid: 1953228
Closes: odoo#8291
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>