[FIX] web_enterprise: detach home menu before rendering Calendar view
Before this commit, the home menu was unmount after the the calendar was
appended to the DOM. So when on_attach_callback for the Calendar view is
called the height calculated is wrong.

DOM when on_attach_callback is called :
+---------------------------------------------+
|           NAV BAR                           |
+---------------------------------------------+
|                                             |
|           HOME MENU                         |
|                                             |
+---------------------------------------------+
|                                             |
|           CALENDAR                          |
|                                             |
+---------------------------------------------+

After this commit, the unmount of home menu is called before. And now
when the on_attach_callback is trigger the home menu is not present
anymore.

Steps to reproduce:
* Open Odoo on a SmallScreen ("Mobile" mode)
* Select Calendar app (BUG calendar is too small)