[IMP] social_linkedin: add streams and statistics
Purpose
=======
Add full support for LinkedIn. We want to be able to read the LinkedIn
posts in the feed view, to read/create comments on posts,
and to retrieve statistics on our posts and on our accounts.

Specifications
==============
Allow the user to see the feed of his LinkedIn company page in the feed
view of social (as we have now the permission from LinkedIn to read the
company pages feed).

Now, we link company pages instead of personal account, because we can
not read the feed of the LinkedIn personal accounts and we want to be
coherent with other medias.

Allow the user to see the comments on the LinkedIn posts, to create
comment and to load replies on comment.

With the LinkedIn API, we can not get comments under a posts and all
replies the the comments in the same request. So, when we click on
"Load X replies", we request the replies to the LinkedIn API before
showing them.

We can not retrieve / send image in comment and we can not edit comment,
so we added a method that allow us to hide the image button near the
comment.

Task 2156000
PR #7550
See odoo/upgrade/pull/782