[IMP] industry_fsm_sale: SO will only generated when product will be added
Currently when clicking on 'product' state button from task view
it generates a sales order but actually should be generated when the
product is added.

so generate an order when the first product will be added.
As we need to set the customer on task, raise userError when when
click on 'product' state button instead of when first product is added.

closes odoo/enterprise#7721

Taskid: 2171538
Closes: #7721
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>