[FIX] web_studio: duplicate report=>copy inherited translation
When in studio you duplicate a report with a view like B:

```
<record name="A">
    <div>text 1</div>
</record>
<record name="B" inherit_id="A"/>
    <xpath expr="." position="inside">
        <div>text 2</div>
    </xpath>
</record>
```

Studio create duplicate B (called B_copy_1) and overwrite it with the
combined heritage (to avoid needing to also duplicate A).

But only the translations of B are duplicated to the new view, so
for the example, translation of `text 2` will be kept, but of `text 1`
will be lost.

So we need to also copy the translations of all the views used to get
the combined view.

Additioanl fixes:

1. if `<t t-name/>` of copied view was root tag, it was never updated to
   the copied view name.

2. the new view containing the combined views was copied before method
   `read_combined` was called to fill it, so it view could possibly be
   applied two times.

Without the changes, the added test failed with:

    self.assertEqual(copy_view.arch, combined_arch)
    AssertionError: '<div>a_<div>ab</div><div>a_</div>aba<div>ab</div><div>ab</div></div>'
                 != '<div>a_<div>ab</div><div>a_</div>aba<div>ab</div></div>'

without this assert (fixed by additional fix 2) the test fails with

    self.assertEqual(set(translations.mapped('src')), set(['a_', 'ab', 'aba']))
    AssertionError: Items in the second set but not the first: 'aba' 'a_'

opw-2154836
closes #7695

closes odoo/enterprise#7756

X-original-commit: 1b4cef4773537b4d7d6a647b2d05de9f7a9c267d
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>