[FIX] web_studio: fix web_studio_tour
Before this commit, the web_studio_tour would break randomly.
This was due to the fact that the tour did not wait for the app creator
to be present in the dom before going to the next step: click on the
button "Close" (Studio).
The fact that the tour would break when the app creator is not there
and the button "Close" clicked is perfectly consistent and coherent
with the specs.
Thus this fix just ensures that the app creator is in the dom
before clicking on "Close" (this is the situation we want to test).

In the same way, we wait for the home menu to be mounted
before triggering the keyboard event 'Escape', this in
order to be sure that the event is handled correctly and the tour
is making what is expected of it.

closes odoo/enterprise#7749

Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>