[IMP] hr_mobile,web_mobile: keep mobile app account details in sync
Before this commit, the mobile device wasn't notified on user's profile
modifications, like updated username, account picture, etc.. This
creates a data desynchronization between the user's profile on the
server and in the mobile app.

This commit adds the necessary native bindings and hooks into both the
startup of the webclient and when the user changes his own preferences
to trigger an update in the mobile application.

Task ID: 2176479

closes odoo/enterprise#7754

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>