[FIX] l10n_de_reports: sollhaben column when an invoice have negative lines
In the case of an invoice with a negative line, the 'sollhaben' column
in the report wasn't correctly filled. Note that, this column should
indicate if the line is debit or credit. Before this commit, this column
was computed based only on the sign of the first line of the move of the
invoice, creating an error if this line was a negative one in an invoice.

opw-2153973

closes odoo/enterprise#7789

X-original-commit: f01edbe4776f6f450e65a7af01e48b8694a1c23a
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>