[FIX] pos_iot: properly close off POS tour
There was a possible race condition if the last step of a tour
performed an action triggering asynchronous side-effects: depending
how fast the headless runner notices it's possible that it starts the
browser's teardown before the async operation has completed, leading
to the resuming callback executing in a partially corrupted state.

In order to avoid this race, the tour runner has been modified to
"blank out" default actions for the last tour step (that is if the
step is left completely blank it's replaced by a no-op instead of the
usual click or value-setting).

For the `payment_terminals_tour` this means the last step of
validating the payment would be skipped, very reliably (instead of
always going through after some time and the next step possibly
generating an error in the browser's cleanup).

Add a step waiting for the receipt page to get displayed instead.

closes odoo/enterprise#7813

X-original-commit: 77000f7c480f9b59b7de3e2f340fdbaf0b94b401
Related: odoo/odoo#43588
Signed-off-by: Xavier Morel (xmo) <xmo@odoo.com>