[IMP] project_enterprise: remove the gantt view from project tasks
Currently Gantt view seems superfluous. it is quite limited and
undermines the use of the Planning app.

so remove the gantt view from tasks view switcher.

The planned_date_begin and planned_date_end appear two time on
the form view of the task when we directly install the industry_fsm
due to the equal view priority of the view.(if we install
project_enterprise first then install the fsm then it works fine
because on fsm task form view is insherited from hr_timesheet not from
project_enterprise)

from project_enterprise this two fields are added after date_deadline
and in industry_fsm this whole group will be replaced but due to same
priority this portion is not replaced so set the priority on
project_enterprise view.

closes odoo/enterprise#7793

Task-id: 2172337
Closes: #7793
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>