[IMP] industry_fsm_stock : use user's default warehouse
Purpose
======
Our current integration between FSM and Stock works well for basic flows, but quickly shows limitations. We don't handle warehouse by employee. Use-case: each worker has a dedicated van with a dedicated stock. I want to track the material used for each intervention to make sure that nothing is stolen.

+ This is also useful in usecases outside of FSM. Example: I have 3 shops (= 3 warehouses), each employee works in a different shop, the default warehouse set on the Sales Order should be the default warehouse of the salesperson. So this should be done sale_stock (and not in FSM)

Specifications
===========
Add a 'Default Warehouse' field on the user only visible if :

- sale_stock is installed
- the 'multi-warehouses' feature is enabled
- domain should only include warehouses which belong to a company the user has access to
- not required

The field should be editable by a user on his own profile and by the admin on all users.

When creating a FSM task from scratch, adding some products creates a SO, the warehouse on the SO should be the default warehouse of the user set on the FSM task.

On SO creation,  set the 'default warehouse' selected on the user assigned to the SO. If there isn't any 'default warehouse' set on the user, fallback on the current behavior (warehouse with
lowest sequence)

The field on the res.user should be a property field to allow multicompany usage.

Task : 2166382

closes odoo/enterprise#8023

Related: odoo/odoo#44257
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>