[FIX] inter_company_rules: journal of counterpart invoice
opw-2171175
The type of the journal was the oposite of what it needed to be (sale
instead of purchase and vice versa) on the invoice created in the other
company.

The default_get for the journal was executed in the new() function,
without the default_type set in the context. This cause the value for
journal_id to be set in the create values, and it wasn't recomputed at
the real create.
To fix this, we have to add the default_type in the context when new()
is executed.

closes odoo/enterprise#7805

X-original-commit: 5812f70dc9bc63deeb120096df03a8ff9f16e48d
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>