[FIX] voip: no dial number when hitting Enter
Before this commit, during a call, when dialing number with click and
then pressing Enter, it was dialing the latest number again. So if we
wrote 12 and press Enter, it will add a 2 and output 122.

This happened because the focus was kept on dial number button, so
pressing Enter still triggered a click on this dial number button.

This commit fixes the issue by setting the focus on to keypad input
after click on dial number. That way, pressing Enter should behave
as expected.

Task-Id 2071310

closes https://github.com/odoo/enterprise/pull/6934

closes odoo/enterprise#6934

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>