[IMP] web: search panel counter changes
This is the companion of a community commit named the same way (see PR
https://github.com/odoo/odoo/pull/44490) bringing some changes in the
search panel.
The main change brought is tha the record counts are now also available
for the fields with select="one" (if not disabled explicitely).
The present commit adapt the related code in documents to the changes.

TASK-ID : 2166814

closes odoo/enterprise#8113

Related: odoo/odoo#44490
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Co-authored-by: Mathieu Duckerts-Antoine <dam@odoo.com>
Co-authored-by: Alexis Lacroix <laa@odoo.com>
Co-authored-by: Julien Mougenot <jum@odoo.com>