[IMP] mrp_workorder, quality_mrp_workorder: link check to production
It would allow to do an easier tracability between check and production.
The domain added on check_ids is used in order to manipulate only check
not linked to a workorder on the production order. Since the check on
the workorder are the steps and should not be processed the same way
than the classic ones.

Task: 31065