[FIX] web_enterprise: notebook's tabs styling on mobile
On mobile, in form view with a notebook (in every app) the line just
below the tabs is too short and when a tab is selected, the line under
this tab doesn't disappear. Also, this line should go from one side to
the other.

This commit fixes those issues by introducing a wrapper element that
handles the scrolling of the tabs. This is required since applying an
`overflow` on an element creates a stacking context preventing the
children from stacking over their parent (and over its bottom border to
hide it).

This commit is also an opportunity to clean up some redundant styles for
those tabs.

Task ID: 2150993

closes odoo/enterprise#7866

X-original-commit: e07aafbe6fee9e568274c0e157645925dc1f6878
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>
Co-authored-by: Dharmraj Jhala <dja@odoo.com>