[FIX] web_grid: text-overflow ellipsis not applied on row header
Before this commit, on web_grid view the row header has attribute
text-overflow ellipsis but it dosn't work because the rule display flex
is applied for all descendant. So the text-overflow is not effective.

After this commit, the display rule is applied to the direct child
and not all descendant.

DOM of grid view:

Before
...
  tbody
    tr
      th
        div display:flex;
          div display:flex;text-overflow:ellipsis;

After
...
  tbody
    tr
      th
        div display:flex;
          div text-overflow:ellipsis;

Ref:
https://www.w3.org/TR/css-overflow-3/#propdef-text-overflow

Step to reproduce:
* open Timesheet app (BUG)

closes odoo/enterprise#7858

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>