[FIX] web_enterprise: notebook tab headers container on desktop
This commit fixes an issue introduced in the commit odoo/enterprise@e07aafbe6fee9e568274c0e157645925dc1f6878
where the notebook's headers container was added only on mobile instead
of all the time. This commit removes the code that will be reintroduced
in community (in the original _renderTagNotebook method).

closes odoo/enterprise#7906

X-original-commit: 1936317b90ee13ddb335f0dd1010dc23d174c1e1
Related: odoo/odoo#43843
Signed-off-by: Pierre Paridans <pparidans@users.noreply.github.com>